Refactor taint analysis for language support
- Introduced `TaintTables` struct to encapsulate language-specific taint configurations, including source idioms, coercions, and sink classifications.
- Updated JavaScript, Python, and Go language modules to utilize `TaintTables` for managing taint sources and sinks.
- Refactored taint detection logic to eliminate language-specific enums and streamline the flow of taint analysis.
- Enhanced modularity by making `ast`, `flow`, `sinks`, `sources`, and `sanitizers` modules public within the `taint` module.
- Removed redundant code and improved clarity in the taint detection process across different languages.
